(Refactor)-페이지 관심사 분리 리펙터링
